Role enum moved to enum package
