+++ src/main/java/com/rezztoran/rezztoranbe/enums/Role.java
@@ -1,4 +1,4 @@
-package com.rezztoran.rezztoranbe.model;
+package com.rezztoran.rezztoranbe.enums;
 
 public enum Role {
     ADMIN, USER, RESTAURANT_ADMIN
